+++ HorizonRadio.UI/Views/QuickPlayDialogView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace HorizonRadio.UI.Views;
+
+public partial class QuickPlayDialogView : UserControl
+{
+    public QuickPlayDialogView()
+    {
+        InitializeComponent();
+    }
+}